+++ src/components/index.js
@@ -3,5 +3,5 @@ export { Footer } from "./footer/Footer";
 export { Navigation } from "./navigation/Navigation";
 export { Sidebar } from "./sidebar/Sidebar";
 export { AppHeading } from "./appHeading/AppHeading";
-export { FormInput } from "./FormInput/FormInput";
-export { FormButton } from "./FormButton/FormButton";
+export { FormInput } from "./formInput/FormInput";
+export { FormButton } from "./formButton/FormButton";
